factor accesstoken runtime-dependency wrapper out
